Renamed all relevant functions. No more fun for anyone
