+++ hatServer/static/partials/build.html
@@ -1,5 +1,5 @@
 ﻿<div class="container">
   <h1>
-    <button type="submit" ng-click="buildBitches()" class="btn btn-default">BUILD THEM BITCHES!</button>
+    <button type="submit" ng-click="buildStudents()" class="btn btn-default">BUILD THEM Students!</button>
   </h1>
 </div>
