+++ hatServer/static/partials/sort.html
@@ -1,5 +1,5 @@
 ﻿<div class="container">
   <h1>
-    <button type="submit" ng-click="sortBitches()" class="btn btn-default">SORT THEM BITCHES!</button>
+    <button type="submit" ng-click="sortStudents()" class="btn btn-default">SORT THEM Students!</button>
   </h1>
 </div>
